Refactor pod status throttling configuration and test layout.
Move pod status/event throttling logic into a dedicated cache module, eliminate duplicated default definitions and package-level mutable throttle state, and expand focused unit coverage for configuration fallback plus boundary behavior.
