+++ cmd/scheduler/app/options/options.go
@@ -46,20 +46,21 @@ const (
 	defaultBurst = 2000
 
 	// Default parameters to control the number of feasible nodes to find and score
-	defaultMinPercentageOfNodesToFind     = 5
-	defaultMinNodesToFind                 = 100
-	defaultPercentageOfNodesToFind        = 0
-	defaultLockObjectNamespace            = "volcano-system"
-	defaultNodeWorkers                    = 20
-	defaultPodStatusLowPressureThreshold  = 500
-	defaultPodStatusHighPressureThreshold = 2000
-	defaultPodStatusLowPressureInterval   = 0 * time.Second
-	defaultPodStatusMidPressureInterval   = 120 * time.Second
-	defaultPodStatusHighPressureInterval  = 300 * time.Second
-	defaultPodStatusForceSyncInterval     = 10 * time.Minute
-	defaultPodEventLowPressureInterval    = 0 * time.Second
-	defaultPodEventMidPressureInterval    = 60 * time.Second
-	defaultPodEventHighPressureInterval   = 120 * time.Second
+	defaultMinPercentageOfNodesToFind = 5
+	defaultMinNodesToFind             = 100
+	defaultPercentageOfNodesToFind    = 0
+	defaultLockObjectNamespace        = "volcano-system"
+	defaultNodeWorkers                = 20
+	// Exported defaults for pod status/event adaptive throttling.
+	DefaultPodStatusLowPressureThreshold  = 500
+	DefaultPodStatusHighPressureThreshold = 2000
+	DefaultPodStatusLowPressureInterval   = 0 * time.Second
+	DefaultPodStatusMidPressureInterval   = 120 * time.Second
+	DefaultPodStatusHighPressureInterval  = 300 * time.Second
+	DefaultPodStatusForceSyncInterval     = 10 * time.Minute
+	DefaultPodEventLowPressureInterval    = 0 * time.Second
+	DefaultPodEventMidPressureInterval    = 60 * time.Second
+	DefaultPodEventHighPressureInterval   = 120 * time.Second
 )
 
 var (
@@ -197,15 +198,15 @@ func (s *ServerOption) AddFlags(fs *pflag.FlagSet) {
 	fs.BoolVar(&s.DisableDefaultSchedulerConfig, "disable-default-scheduler-config", false, "The flag indicates whether the scheduler should avoid using the default configuration if the provided scheduler configuration is invalid.")
 	fs.StringVar(&s.ShardingMode, "scheduler-sharding-mode", util.NoneShardingMode, "The node sharding mode for scheduling, none(default)|hard|soft mode is supported")
 	fs.StringVar(&s.ShardName, "scheduler-sharding-name", defaultShardName, "The name of shard used for this scheduler")
-	fs.IntVar(&s.PodStatusLowPressureThreshold, "pod-status-low-pressure-threshold", defaultPodStatusLowPressureThreshold, "Pending pod count threshold below which pod status throttling is disabled (this is pod count, not node count)")
-	fs.IntVar(&s.PodStatusHighPressureThreshold, "pod-status-high-pressure-threshold", defaultPodStatusHighPressureThreshold, "Pending pod count threshold above which high-pressure throttling strategy is used (this is pod count, not node count)")
-	fs.DurationVar(&s.PodStatusLowPressureInterval, "pod-status-low-pressure-interval", defaultPodStatusLowPressureInterval, "Throttle interval for repeated unschedulable pod status updates under low pressure")
-	fs.DurationVar(&s.PodStatusMidPressureInterval, "pod-status-mid-pressure-interval", defaultPodStatusMidPressureInterval, "Throttle interval for repeated unschedulable pod status updates under medium pressure")
-	fs.DurationVar(&s.PodStatusHighPressureInterval, "pod-status-high-pressure-interval", defaultPodStatusHighPressureInterval, "Throttle interval for repeated unschedulable pod status updates under high pressure")
-	fs.DurationVar(&s.PodStatusForceSyncInterval, "pod-status-force-sync-interval", defaultPodStatusForceSyncInterval, "Maximum staleness before forcing a pod status sync even when repeated unschedulable updates are throttled")
-	fs.DurationVar(&s.PodEventLowPressureInterval, "pod-event-low-pressure-interval", defaultPodEventLowPressureInterval, "Throttle interval for repeated unschedulable pod events under low pressure")
-	fs.DurationVar(&s.PodEventMidPressureInterval, "pod-event-mid-pressure-interval", defaultPodEventMidPressureInterval, "Throttle interval for repeated unschedulable pod events under medium pressure")
-	fs.DurationVar(&s.PodEventHighPressureInterval, "pod-event-high-pressure-interval", defaultPodEventHighPressureInterval, "Throttle interval for repeated unschedulable pod events under high pressure")
+	fs.IntVar(&s.PodStatusLowPressureThreshold, "pod-status-low-pressure-threshold", DefaultPodStatusLowPressureThreshold, "Pending pod count threshold below which pod status throttling is disabled (this is pod count, not node count)")
+	fs.IntVar(&s.PodStatusHighPressureThreshold, "pod-status-high-pressure-threshold", DefaultPodStatusHighPressureThreshold, "Pending pod count threshold above which high-pressure throttling strategy is used (this is pod count, not node count)")
+	fs.DurationVar(&s.PodStatusLowPressureInterval, "pod-status-low-pressure-interval", DefaultPodStatusLowPressureInterval, "Throttle interval for repeated unschedulable pod status updates under low pressure")
+	fs.DurationVar(&s.PodStatusMidPressureInterval, "pod-status-mid-pressure-interval", DefaultPodStatusMidPressureInterval, "Throttle interval for repeated unschedulable pod status updates under medium pressure")
+	fs.DurationVar(&s.PodStatusHighPressureInterval, "pod-status-high-pressure-interval", DefaultPodStatusHighPressureInterval, "Throttle interval for repeated unschedulable pod status updates under high pressure")
+	fs.DurationVar(&s.PodStatusForceSyncInterval, "pod-status-force-sync-interval", DefaultPodStatusForceSyncInterval, "Maximum staleness before forcing a pod status sync even when repeated unschedulable updates are throttled")
+	fs.DurationVar(&s.PodEventLowPressureInterval, "pod-event-low-pressure-interval", DefaultPodEventLowPressureInterval, "Throttle interval for repeated unschedulable pod events under low pressure")
+	fs.DurationVar(&s.PodEventMidPressureInterval, "pod-event-mid-pressure-interval", DefaultPodEventMidPressureInterval, "Throttle interval for repeated unschedulable pod events under medium pressure")
+	fs.DurationVar(&s.PodEventHighPressureInterval, "pod-event-high-pressure-interval", DefaultPodEventHighPressureInterval, "Throttle interval for repeated unschedulable pod events under high pressure")
 }
 
 // CheckOptionOrDie check leader election flag when LeaderElection is enabled.
