+++ cmd/scheduler/app/options/options_test.go
@@ -87,15 +87,15 @@ func TestAddFlags(t *testing.T) {
 		ShardingMode:                   commonutil.NoneShardingMode,
 		ShardName:                      defaultSchedulerName,
 		ResourceSyncTimeout:            60 * time.Second,
-		PodStatusLowPressureThreshold:  defaultPodStatusLowPressureThreshold,
-		PodStatusHighPressureThreshold: defaultPodStatusHighPressureThreshold,
-		PodStatusLowPressureInterval:   defaultPodStatusLowPressureInterval,
-		PodStatusMidPressureInterval:   defaultPodStatusMidPressureInterval,
-		PodStatusHighPressureInterval:  defaultPodStatusHighPressureInterval,
-		PodStatusForceSyncInterval:     defaultPodStatusForceSyncInterval,
-		PodEventLowPressureInterval:    defaultPodEventLowPressureInterval,
-		PodEventMidPressureInterval:    defaultPodEventMidPressureInterval,
-		PodEventHighPressureInterval:   defaultPodEventHighPressureInterval,
+		PodStatusLowPressureThreshold:  DefaultPodStatusLowPressureThreshold,
+		PodStatusHighPressureThreshold: DefaultPodStatusHighPressureThreshold,
+		PodStatusLowPressureInterval:   DefaultPodStatusLowPressureInterval,
+		PodStatusMidPressureInterval:   DefaultPodStatusMidPressureInterval,
+		PodStatusHighPressureInterval:  DefaultPodStatusHighPressureInterval,
+		PodStatusForceSyncInterval:     DefaultPodStatusForceSyncInterval,
+		PodEventLowPressureInterval:    DefaultPodEventLowPressureInterval,
+		PodEventMidPressureInterval:    DefaultPodEventMidPressureInterval,
+		PodEventHighPressureInterval:   DefaultPodEventHighPressureInterval,
 	}
 	expectedFeatureGates := map[featuregate.Feature]bool{
 		features.PodDisruptionBudgetsSupport: false,
